# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Restful API of Artshow",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000/"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/artists": {
      "get": {
        "tags": [
          "Artist"
        ],
        "description": "List all artists",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/artist/{id}": {
      "get": {
        "tags": [
          "Artist"
        ],
        "description": "Get the details of an artist by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Artist"
        ],
        "description": "update the details of an artist by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Artist"
              }
            }
          }
        }
      }
    },
    "/artworks": {
      "get": {
        "tags": [
          "Artwork"
        ],
        "description": "List all artworks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/artwork/{id}": {
      "get": {
        "tags": [
          "Artwork"
        ],
        "description": "Get the details of an artwork by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully obtained.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 15
                    },
                    "url": {
                      "type": "string",
                      "example": "https://cdn.pixabay.com/photo/2023/02/28/15/58/off-grid-living-7821244__340.jpg"
                    },
                    "author_id": {
                      "type": "number",
                      "example": 3
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 15
                    },
                    "url": {
                      "type": "string",
                      "example": "https://cdn.pixabay.com/photo/2023/02/28/15/58/off-grid-living-7821244__340.jpg"
                    },
                    "author_id": {
                      "type": "number",
                      "example": 3
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Artwork"
        ],
        "description": "update the details of an artwork by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Artwork"
              }
            }
          }
        }
      }
    },
    "/artworks/by/author/{authorId}": {
      "get": {
        "tags": [
          "Artwork"
        ],
        "description": "list all artworks by authorId",
        "parameters": [
          {
            "name": "authorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/upload-photos": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Artist": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "name"
          },
          "bio": {
            "type": "string",
            "example": "I feel lucky"
          },
          "id": {
            "type": "number",
            "example": 123
          },
          "avatar": {
            "type": "string",
            "example": "http://example.jpg"
          }
        },
        "xml": {
          "name": "Artist"
        }
      },
      "Artwork": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": ""
          },
          "author_id": {
            "type": "string",
            "example": ""
          },
          "id": {
            "type": "number",
            "example": 12
          }
        },
        "xml": {
          "name": "Artwork"
        }
      }
    }
  }
}